Move resizing functions to printer/mod.rs
